requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.1
django-chartit==0.2.9
django-chartjs==1.2
django-jquery==3.1.0
djangorestframework==3.7.7
flake8==3.5.0
idna==2.6
mccabe==0.6.1
pew==1.1.2
pipenv==10.0.0
pkg-resources==0.0.0
pycodestyle==2.3.1
pyflakes==1.6.0
pymodbus==1.4.0
pyserial==3.4
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
djangorestframework==3.7.7
django-bootstrap3==9.1.0
django-chartit==0.2.9
chartjs==1.2
django-chartit==0.2.9
